Fix implicit conversion issues with Xcode 7.3

+++ cocos/base/CCVector.h
@@ -288,7 +288,7 @@ public:
     {
         if (!_data.empty())
         {
-            ssize_t randIdx = RandomHelper::random_int<int>(0,_data.size()-1);
+            ssize_t randIdx = RandomHelper::random_int<int>(0, static_cast<int>(_data.size()) - 1);
             return *(_data.begin() + randIdx);
         }
         return nullptr;
